refactor(articles): extract content generation boundaries
